Update rules.md
I just took the priority hover description and modified it a little to be a bit more semantic. I thought maybe calling 10 the "lowest" might be confusing for some. So I changed the semantics to be "more" or "less" important.

I took some liberty on it, so please let me know if you don't want it changed!